solved exercise 2
